chore: fix paths after failed merge
Things have moved, so paths must be adapted afterwards.

For the `addFormHiddenField` method, which has been moved to demosplan-core,
we decided to duplicate it here (as long as "classic" forms have to be supported),
because the only alternative approach that makes sense here would be to introduce
the possibility to introduce some kind of hooks in the validation process, which
is out of scope for dpValidate as of now.