fix(refs T31971): replace reserved chars in filenames
As the tus endpoint breaks with 400 when uploading files with
certain characters within them, the fileName is cleaned from
those malfunctioning characters before adding it to the Uppy store
(and eventually uploading them, as auto resume is activated).

The characters to be replaced were chosen based on a manual test
that i conducted - while i could not produce a file containing
a forward slash on my linux machine, all other characters have
proven to break the endpoint. Maybe i missed a character, then
it will have to be added to the list.

The fix revealed another bug - file names containing a colon :
now broke the frontend, because the colon is used as a separator
in getFileInfo. As the conversion (from object to fileString and
vice versa) is not needed but is a remainder from the Uppy refactoring
(as far as i could understand it), it is dropped altogether, and the
file object is directly passed into DpUploadFile.