feat(refs: T30974): Add design tokens for boxShadow, color, fontSize and rounded (#108)
This commit establishes a working set of design tokens within demosplan-ui. It adds
tokens for color, boxShadow, fontSize and rounded, as well as an internal `Status` token.

New Hex Colors:
- Palettes are supplemented with the new color palette teal
- The magenta palette has been refined to also have 7 steps, values have been adapted to form a more regular curve
- The value of color.palette.yellow.dark-1 has been adapted to be slightly darker, to better match the yellow palette curve
- The value of color.palette.red.base has been tweaked to be slightly less aggressive

New Brand Tokens:
3 pairs of brand tokens are introduced: main, alt and hightlight - each with their contrast counterpart holding a color value that ensures accessibility when a pair is used as foreground/background color.

The following alias token families have been added to demosplan-ui:
- interactive
- message
- status
- text
- selection
- border
- background

A more detailed, yet incomplete, description of each token can be found in the `tokens.color.stories.mdx` file. A thorough description can be found in the `$description` field of each token within the token definition in `color.ui.json`.

New Scss variables structure:
When running `yarn build-tokens`, Scss variables for colors do not contain the "domain" part (ui, palette eg...). 
We discussed this with the team, and agreed on the variables being expressive enough to drop the "domain" 
part and shorten variable length a little.

New File Structure:
Palette, Data, Brand, and UI tokens are defined within their own json files. 
The generated Scss files also follow that structure.
For internal tokens, a new `_.json` File is created.

Documentation:
To document the basics, a tokens overview and a color tokens article are established. 
The latter still leaves out some UI tokens that still have to be documented. 
As it is planned to generate color tokens on an automated base, this is left out intentionally for now.