feat(refs T21517) Add subtle button variant + icons (#259)
Apart from the existing button styles, a more subtle variant is needed.
To reduce confusion, the default button variant, which went without a name
before (as its styles were directly attached to the .btn class) is named "solid".

This also adds the ability of passing icons to the button component,
dropping the slot in advance.

For changed styling, also see demos-europe/demosplan-core#1436

In the future there may be other button variants ("link" which may appear
like a text link). This variant, however, focuses on the reduced level of
importance and visibility.