F t32983 move generic components to demosplan  UI (#473)
* feat(T32983): add addon wrapper to demosplanui

also rename it to DpAddonWrapper. This happens for two reasons.

- This is a really generic component and hence we should handle this component like other generic ones too, by addin them to demosplan-ui
- There is an addon that needs to load an addon. Because I want to avoid copying the same file into multiple repositories, this solution made the most sense to me.

* feat(T32983): move dp badge to demosplan ui

this generic component is needed in multiple repositories hence should live in demosplan-ui

* feat(T32983): this component will be shared with another addon

hence this component is generic enough to qualify for demosplan-ui

* feat(T32983): add components to index.js

so these files can be imported out of the demosplan-ui npm package

* feat(T32983): fix missing prefix

* feat(T32983): remove DpInsertableRecommendation from demosplan-ui

* feat(T32983): tweak comments and fix typos

* feat(T32983): remove DpAddonWrapper from ui

this component can live in the core because we will refactor the tag based recommendations in the recommendation modal to be a tab. This way we prevent the need to have an addon load an addon (which sounded good at first, but I didn't think about the possible resulting confusion which would come along)

* feat(T32983): add DpBadge boilerplate story

- with minimal configuration and explanation

* feat: fix some docs details

* feat(T32983): add changelog entry

---------

Co-authored-by: Velimir Müller <velimir@demos-deutschland.de>
Co-authored-by: Hans Spieß <spiess@demos-deutschland.de>