feat: Use alias "~/" in components (#515)
With #513, we introduced the internal alias `~` (which points to `./src`), 
along with an index.ts in every component folder. This change adds index files
for all components and uses the `~` where appropriate.

Also, components that lived within DpDataTable directory but are not direct
dependencies of DpDataTable are now moved to their own folders:
- DpDataTableExtended
- DpColumnSelector

Also, due to the introduction of the index files, within stories we can now
import components from the same folder with `./`.