feat(refs T20094): Add spacing tokens to tailwind (#668)
This change consolidates spacing tokens and makes them available to Tailwind.

Basically, the already existing space values that proved to be useful are preserved 
(but renamed in a tailwind-like manner), the larger steps are recalibrated using 
a geometric progression. The weird factor of 1.371 is chosen because with this factor, 
w-14 turns out to be nearly 600px and there is a smooth transition
from the linear progression in steps 1-8.

This PR also brings a jsx Table component to Storybook, along with a babel preset-react. 
This needs to happen because markdown tables are somewhat broken in SB7. 
All token doc tables are equipped with that new table, now.