Fix(T35880): improve handling of non json responses (#690)
This PR handles cases where the response does not contain content and cannot be parsed as JSON.

In Bug Ticket T35884, @gruenbergerdemos has experienced a 'Bad Request' issue. Method names are case-sensitive and should be in uppercase, so we receive a 400 Bad Request error if they are in lowercase. Fixed it in this PR as well.