fix: always serialize request params
When moving from axios to fetch, it was overlooked
that axios serializes params by default, and we only
do that occasionally. This leads to the params being
dropped in the fetch request whenever "serialize: true"
is set via the options arg.

To overcome this, and as we always need to serialize params
within get requests, the request params are always appended
as stringified value to the url. As the "url" argument may
already have params appended, this needs to be handled as well.

The "payload" variable has been renamed into "fetchOptions"
to level up readability while distinguishing it from the
"options" variable still being used in dpApi.

Also, "options" and "params" has been removed from the request
payload as the fetch method does not use these values.

See https://developer.mozilla.org/en-US/docs/Web/API/fetch