Add button styles (#964)
This change adds styling via Tailwindcss to the DpButton component. 
To make this work, several things had to be done:

- new tokens are added to style secondary and warning buttons. to not have to redefine 
  those for backgroundColor and borderColor, they're made available for all color modules
- tailwind lineHeight is configured with our spacing tokens
- rounded-button is available to style/override button radius
- `bg-busy` and `animate-busy` are available to style a busy animation

There is yet one unsolved task: with the current way of configuring Tailwindcss with 
`var(--a, <value>)` it is not possible for Tailwind to generate alpha variants for colors. 
So, instead of referencing the token values, the respective hex colors had to be 
hardcoded in the `focus-visible:outline-[#B20000]/50` outline color definitions. 
As the proposed solution involves reworking the whole Tailwindcss config and 
StyleDictionary transform, i did not approach that within the current change.