fix: restore installability of package (#983)
Due to several things, the yarn.lock and package.json
were out of sync, so that vue-loader was not installed
anymore by yarn 1 and the build failed consequently.

This change restores the state of the package.json
before bumping Yarn to 4, and it additionally removes
"@webpack-cli/generators" which seems to be a remnant of
bootstrapping the project build.

Also, this change bumps several dependencies to their latest
resolvable minor versions as i deleted yarn.lock before
running `yarn install` to have a clean start, again.

Lastly, this change fixes string-width and wrap-ansi at specific
resolutions, to overcome the "string-width-error" once more.