feat(refs DPLAN-12968): introduce condensed search field (1133)
* feat(refs DPLAN-12968, #AB21840): introduce condensed search field

- the search button is changed to an icon-only outline button using the magnifying
glass icon and is attached to the right side of the search field
- before, a primary button was used, but in most cases, the search action is not
the primary action in a page, so we decided to use the outline button instead

* feat(refs DPLAN-12968, #AB21840): restore left button border

if we remove the left border of the button and we use the input
border instead, it will not have the desired color, so a different
solution is implemented in demosplan-core

* docs(refs DPLAN-12968, #AB21840): add changelog entry

* feat(refs DPLAN-12968, #AB21840): allow passing additional buttons

in some cases, we may want to pass additional buttons to DpResettableInput,
e.g. when using it with DpSearchField we sometimes need a 'settings' button

* feat(refs DPLAN-12968, #AB21840): allow passing additional buttons

in some cases, we may want to pass additional buttons to DpResettableInput,
e.g. when using it with DpSearchField we sometimes need a 'settings' button

* refactor(refs DPLAN-12968, #AB21840): remove unnecessary argument

* style(refs DPLAN-12968, #AB21840): remove unused method

* test(refs DPLAN-12968, #AB21840): add unit test for DpSearchField

* feat(refs DPLAN-12968, #AB21840): remove obsolete prop + class

after removing the right border from the search input in c1cf03c9335f65591e3a368aed4d0e6e9154712e,
we don't need to set the css class in DpInput anymore

* feat(refs DPLAN-12968, #AB21840): remove filled gear icon

for now, we don't need it anymore

* feat(refs DPLAN-12968, #AB21840): remove unnecessary line break

* style(refs DPLAN-12968, #AB21840): use inline-flex instead of align-top + comment

to be able to remove the comment between dp-resettable input and dp-button
as well as the align-top class on dp-button, we now use inline-flex on the outer
span

* style(refs DPLAN-12968, #AB21840): add missing category