fix(refs DPLAN-15151): Propperly emit checkboxGroup events (1214)
hence the new value where not updated when the emit got fired, we had to set the updated value manually to the emit payload.
And because we listen to the events straight from the component and not from the parent, we don't have to parent.emit, but can keep it clean and emit directly.

While doing that, I updated to the new syntax and tried to clean the code around a bit