fix(refs DPLAN-15169): fix image upload in DpEditor (1217)
* fix(refs DPLAN-15169): declare emitted events

* fix(refs DPLAN-15169): use correct uppy method

close() does not exist; clear() seems to be what's needed
instead

* style(refs DPLAN-15169): improve code style

* fix(refs DPLAN-15169): correctly access parent element

this.$parent is no longer available in vue 3

* style(refs DPLAN-15169): rename method

this way it seems a littler easier to understand what's going on

* fix(refs DPLAN-15169): correctly pass boolean

* fix(refs DPLAN-15169): fix render function

this allows us to use the vue 3 render function api

* docs(refs DPLAN-15169): add changelog entry

* fix(refs DPLAN-15169): revert change to boolean

the attribute "is enumerated and not Boolean. A value of true or false is mandatory,
and shorthand like <img draggable> is forbidden. The correct usage is
<img draggable="false">""