chore: Use checkResponse after every dpAPi call
With this change the response from checkResponse changes. Instead of returning `response.data`, it now returns `response`. So we have to adjust the structure on all places where we removed checkResponse in here.

This change was necessary because that is the response that would come when checkResponse is not in use. And its much harder to find all occurrences where dpApi is used but not checkResponse than check the occurrences, where we had to remove checkresponse anyway.

And: dpRPC even don't always return a `response.data` but also we get checkResponse out of the box