chore: Migrate storybook to v9 (1310)
with storybook 9, they dropped the webpack support for vue.
That's why we now have to use vite for storybook.

Some of the styles are broken, due to the circumstance that the components get styled from the consuming app. All components with an somehow broken style have to be migrated to tailwind in the End.

Can be tested with:

$ yarn
$ yarn build:tokens
$ yarn build:storybook
$ yarn sb