fix(refs DPLAN-16550): improve DpAutocomplete usability (1358)
* fix(refs DPLAN-16550): use actual input for auto suggest

- by replacing the custom implementation using a div as input box
with an actual input, a lot of UX issues are solved automatically
(e.g. that the cursor is right-aligned when clicking inside the input,
that the cursor can't be moved, that the placeholder disappears when
clicking inside the input)
- keyboard navigation with arrow keys to select suggestions is now
supported
- styling in DpResettableInput is fixed for when content is passed
to the slot (i.e., buttons are displayed next to each other and aligned
to the right)
- reset and search buttons are always displayed

* docs(refs DPLAN-16550): add documentation for DpAutocomplete

* docs(refs DPLAN-16550): add changelog entry

* chore(refs DPLAN-16550): don't expose unused methods

* chore(refs DPLAN-16550): make code style consistent

* fix(refs DPLAN-16550): pass string values to attributes

- attributes like aria-expanded and aria-selected should be set to string values
instead of booleans
- this removes some warnings from the console

* docs(refs DPLAN-16550): update documentation

after refactoring DpAutocomplete, the documentation
needs some updates

* test(refs DPLAN-16550): fix unit test

- since we refactored DpAutocomplete, the test needs to be
updated
- keyboard navigation and accessibility tests are grouped
- code duplication is reduced by extracting functions
- redundant comments are removed
- some selectors needed to be renamed

* refactor(refs DPLAN-16550): pass attrs directly to input

- attrs passed to DpInput (e.g. in DpAutocomplete) are now applied
to the input, not the wrapping div
- if class is passed to DpInput, it is still applied to the div though
- attrs passed to DpInput in DpAutocomplete are aria-haspopup,
aria-controls, aria-expanded, and aria-activedescendant. we don't want
to add them as props to DpInput, because they are not normally needed, but for
DpAutocomplete, we need to pass them to the input element

* test(refs DPLAN-16550): test passing through aria attributes

when aria attributes are passed to DpInput, they should be applied
directly to the input, not the wrapping div

* refactor(refs DPLAN-16550): do not inherit attrs

- when using v-model on DpResettableInput from a component using
composition api, modelvalue is added as an attribute to the
outer div in DpResettableInput, which should not happen
- for now, we set inheritAttrs to false so this does not happen
anymore
- we should refactor DpResettableInput to work with vue 3 v-model
though, instead of using the vue 2 way

* fix(refs DPLAN-16550): prevent hiding of suggestions list after reset

* fix(refs DPLAN-16550): further improve usability

- when a user enters text, then selects a suggestion, then empties
the input via backspace, then enters new text, the options (suggestions)
list is now displayed again
- fetchSuggestions request is debounced, so it's not triggered on every
character input; also, it is not triggered when using backspace
- while searching, a message is displayed, so the user understands that
something is happening