fix(DPLAN-12751): DpTabs: accessibility improvement (#1416)
* fix(DPLAN-12751): This PR improves accessibility in DpTabs and DpBulkEditHeader components by fixing an ARIA issue (aria-required-children):

DpTabs: use a <div> instead of a <ul> for the tabs, since interactive roles should only be applied to appropriate elements.
DpBulkEditHeader: remove role attribute as it is not a menu in the classical sense. A native button provide already the required semantics - so there is no need to add ARIA role attribute